Thu, Mar 24, 2022, 5:39 AM +00:00

+++ filter-less/helpers.c
@@ -92,7 +92,7 @@ void blur(int height, int width, RGBTRIPLE image[height][width])
     }
 
     //find the 4 pixel corners of the image and blur them
-    image[0][0].rgbtBlue = (BYTE) round((copy[0][0].rgbtBlue + copy[0][1].rgbtBlue + copy[1][0].rgbtBlue + copy[1][1].rgbtBlue) / 4.0);
+    image[0][0].rgbtBlue = round((copy[0][0].rgbtBlue + copy[0][1].rgbtBlue + copy[1][0].rgbtBlue + copy[1][1].rgbtBlue) / 4.0);
     image[0][0].rgbtGreen = (BYTE) round((copy[0][0].rgbtGreen + copy[0][1].rgbtGreen + copy[1][0].rgbtGreen + copy[1][1].rgbtGreen) / 4.0);
     image[0][0].rgbtRed = (BYTE) round((copy[0][0].rgbtRed + copy[0][1].rgbtRed + copy[1][0].rgbtRed + copy[1][1].rgbtRed) / 4.0);
